% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chocoparty-package.R
\docType{package}
\name{chocoparty-package}
\alias{chocoparty}
\alias{chocoparty-package}
\title{chocoparty: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
