% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_choco}
\alias{clean_choco}
\title{Clean the columns in the choco data}
\usage{
clean_choco(data)
}
\arguments{
\item{data}{Choco data frame}
}
\value{
Clean choco data frame: number of ingredients separate from ingredients
list. Trim stings, numbers as numeric, ...
}
\description{
Clean the columns in the choco data
}
\examples{
data(choco)
clean_choco(choco)
}
