% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_per_group.R
\name{mean_per_group}
\alias{mean_per_group}
\title{Calculate the mean of a column for data grouped by another column}
\usage{
mean_per_group(data, colGroup, colMean)
}
\arguments{
\item{data}{Data frame}

\item{colGroup}{Name of the column you want to group by}

\item{colMean}{Name of the column you want to calculate the mean of}
}
\value{
A data frame with the group and the mean per group
}
\description{
Calculate the mean of a column for data grouped by another column
}
\examples{
data(choco)
mean_per_group(choco, "company_location", "rating")
}
