% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean.R
\name{plot_mean}
\alias{plot_mean}
\title{Plot the top x value as a function of an other with horizontal bars}
\usage{
plot_mean(data, colGroup, colMean, n_top, image_path)
}
\arguments{
\item{data}{Data frame e.g. Group, Mean per group}

\item{colGroup}{Name of the column to group by}

\item{colMean}{Name of the mean column}

\item{n_top}{Number of rows you want to plot}

\item{image_path}{Path to the image for the points}
}
\value{
A plot
}
\description{
Plot the top x value as a function of an other with horizontal bars
}
\examples{
data(choco)
choco_mean <- mean_per_group(choco, "company_location", "rating")
plot_mean(choco_mean, "company_location", "mean_rating", 5, "/Users/au761602/Downloads/canva-yellow-star-MAFXSVQWSuc.png.webp")
}
