% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_long.R
\name{separate_longer}
\alias{separate_longer}
\title{Separate columns containing delimited info into several rows}
\usage{
separate_longer(data, col_long, other_cols, delim)
}
\arguments{
\item{data}{Data frame}

\item{col_long}{Name of the column to split into multiple rows}

\item{other_cols}{Name of the other columns you want to include in the output table}

\item{delim}{Delimiter}
}
\value{
A long data frame
}
\description{
Separate columns containing delimited info into several rows
}
\examples{
data(choco)
separate_longer(choco, "ingredients", "rating", "-")
}
